+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 - Clear shortened release progress statuses on TTY redraws
 
+- Redraw concurrent TTY progress rows from a stable block anchor.
+
 ### Security
 
 ## [1.2.26] - 2026-08-03